+++ drizzle/0014_modern_nekra.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `sources` ADD `sync_interval_minutes` integer DEFAULT 1440 NOT NULL;